+++ src/Shippable.java
@@ -1,4 +1,4 @@
 public interface Shippable {
     String getName();
-    float getWeight();
+    double getWeight();
 }
